DEV: react native view shot added and implemented, but it only gets the background of the review card
