+++ components/ShareCard/ReviewCard.tsx
@@ -1,0 +1,84 @@
+import { Image, StyleSheet, View, Text } from 'react-native';
+
+import { useTheme } from '@react-navigation/native';
+import { Icon } from 'react-native-elements';
+
+import { ThemedText } from '../ThemedText';
+import { Colors } from '@/constants/Colors';
+
+export const ReviewCard = ({
+  viewRef,
+  selectedPhoto,
+}: {
+  selectedPhoto: string | null | undefined;
+  viewRef: React.RefObject<View>;
+}) => {
+  const styles = themedStyles(Colors);
+
+  return (
+    <View ref={viewRef} style={styles.reviewCard}>
+      <Image
+        source={{ uri: selectedPhoto ? selectedPhoto : '' }} // Replace with actual image
+        style={styles.poster}
+      />
+      <View style={styles.reviewContent}>
+        <Text style={styles.username}>Ghomes</Text>
+        <View style={styles.stars}>
+          {[...Array(5)].map((_, i) => (
+            <Icon
+              key={i}
+              name="star"
+              type="font-awesome"
+              color="#4CAF50"
+              size={18}
+              style={styles.icon}
+            />
+          ))}
+          <Icon name="heart" type="font-awesome" color="#4CAF50" size={18} />
+        </View>
+        <ThemedText style={styles.reviewText}>Cute dog</ThemedText>
+      </View>
+    </View>
+  );
+};
+
+/* eslint-disable react-native/no-unused-styles */
+const themedStyles = (colors: typeof Colors) =>
+  StyleSheet.create({
+    reviewCard: {
+      flexDirection: 'row',
+      backgroundColor: colors().reviewCardBG,
+      borderRadius: 15,
+      padding: 10,
+      alignItems: 'center',
+      width: '90%',
+    },
+    poster: {
+      width: 80,
+      height: 120,
+      borderRadius: 10,
+    },
+    reviewContent: {
+      flex: 1,
+      marginLeft: 10,
+      alignSelf: 'flex-start',
+    },
+    username: {
+      color: colors().text,
+      fontSize: 16,
+      fontWeight: 'bold',
+    },
+    stars: {
+      flexDirection: 'row',
+
+      marginVertical: 5,
+    },
+    reviewText: {
+      color: colors().text,
+      fontSize: 14,
+    },
+    icon: {
+      marginRight: 5,
+    },
+  });
+/* eslint-disable react-native/no-unused-styles */
