+++ package.json
@@ -3,9 +3,9 @@
   "main": "expo-router/entry",
   "version": "1.0.0",
   "scripts": {
-    "start": "expo start",
-    "android": "expo run:android",
-    "ios": "expo run:ios",
+    "start": "watchman watch-del '/Users/Tiago.GOMES/Documents/RN_Social_Sharing' ; watchman watch-project '/Users/Tiago.GOMES/Documents/RN_Social_Sharing' ; expo start",
+    "android": "watchman watch-del '/Users/Tiago.GOMES/Documents/RN_Social_Sharing' ; watchman watch-project '/Users/Tiago.GOMES/Documents/RN_Social_Sharing' ; expo run:android",
+    "ios": "watchman watch-del '/Users/Tiago.GOMES/Documents/RN_Social_Sharing' ; watchman watch-project '/Users/Tiago.GOMES/Documents/RN_Social_Sharing' ; expo run:ios",
     "web": "expo start --web",
     "test": "jest --watchAll",
     "lint": "expo lint",
@@ -42,6 +42,7 @@
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-share": "^12.0.9",
+    "react-native-view-shot": "^4.0.3",
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5"
   },
